% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{dgvm3d.snapshots}
\alias{dgvm3d.snapshots}
\title{Year 2000 snapshot of a LPJ-GUESS simulation}
\format{An object of class \code{list} of length 13.}
\usage{
dgvm3d.snapshots
}
\description{
Simulation with 12 patches at 13 locations (default test gridlist in LPJ-GUESS)
}
\details{
@name dgvm3d.snapshots
 @docType data
 @author Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
 @keywords data
}
\keyword{datasets}

