% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{drawEllipsoid}
\alias{drawEllipsoid}
\title{draw a quadrilateral ellipsoid}
\usage{
drawEllipsoid(rx = 1, ry = 1, rz = 1, ctr = c(0, 0, 0), n = 72,
  col = "green", opacity = opacity)
}
\arguments{
\item{rx}{radius in x direction}

\item{ry}{radius in y direction}

\item{rz}{radius in z direction}

\item{ctr}{3-element vector. center of the ellipsoid}

\item{n}{number of vertices along z-axis and around xy rotation}

\item{col}{the color to use.}

\item{opacity}{the alpha value for the surface}
}
\description{
I copied and modified it from 'Code demonstations' rgl::shapes3d (search with "??ellipsoid" on commandline)
}

