% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{establishPatch}
\alias{establishPatch}
\title{poplulate a patch with its vegetation}
\usage{
establishPatch(vegetation = NULL, radius = 1)
}
\arguments{
\item{vegetation}{the vegetation data.frame}

\item{radius}{the radius used to distribute the vegetation to}
}
\value{
the vegetation data.frame with the positions
}
\description{
Randomly 'plant' the trees in the patch within a given radius.
}
\examples{
dgvm3d.options("default")
stand = initStand(npatch=1)
veg = data.frame(DBH=rep(0.5, 100))
veg$Height    = veg$DBH * 35
veg$Crownarea = veg$DBH * 5
veg$LeafType  = sample(0:1, nrow(veg), replace=TRUE)
veg$ShadeType = sample(0:1, nrow(veg), replace=TRUE)
stand@patches[[1]]@vegetation = establishPatch(veg, stand@hexagon@supp[['inner.radius']])
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

