% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gapless.rank}
\alias{gapless.rank}
\title{ranking without gaps with duplicate values}
\usage{
gapless.rank(x)
}
\arguments{
\item{x}{the input to be ranked}
}
\value{
the ranked data
}
\description{
ranking without gaps with duplicate values
}
\examples{
set.seed(456789)
x = sample(1:20, 12)
x = append(x, x[sample(1:12, 8, replace=TRUE)])
## equal results
gapless.rank(x)
as.numeric(factor(x))
## However, this here keeps the shape
gapless.rank(matrix(x, 4, 5))
as.numeric(factor(matrix(x, 4, 5)))
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

