% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{grass3D}
\alias{grass3D}
\title{Plant the grass on the patch}
\usage{
grass3D(grass = NULL, kind = NULL, offset = c(0, 0, 0), col = "green",
  opacity.threshold = c(0.2, 2), height.scale = 0.1)
}
\arguments{
\item{grass}{vegetation data.frame}

\item{kind}{so far only a hexagon is allowed (TriangBody)}

\item{offset}{the patch offset}

\item{col}{the color to use for grass}

\item{opacity.threshold}{no grass is drawn below the lower values of LAI and full opacity is used above the upper value}

\item{height.scale}{scale the LAI by this factor as height for the hexagon.}
}
\description{
Plant the grass on the patch
}
