% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{initPatch}
\alias{initPatch}
\title{poplulate a patch with its vegetation}
\usage{
initPatch(vegetation = NULL, radius = 1, samples = 3)
}
\arguments{
\item{vegetation}{the vegetation data.frame}

\item{stand}{the stand}

\item{patch.id}{the patch to polulate}

\item{sample}{number of samples to determine the next trees position}
}
\value{
the vegetation data.frame with the positions
}
\description{
poplulate a patch with its vegetation
}
\examples{
stand = initStand(npatch=1)
veg = data.frame(DBH=rep(0.5, 100))
veg$Height    = veg$DBH * 35
veg$Crownarea = veg$DBH * 5
veg$LeafType  = sample(0:1, nrow(veg), replace=TRUE)
veg$ShadeType = sample(0:1, nrow(veg), replace=TRUE)
stand@patches[[1]]@vegetation = initPatch(veg, stand@hexagon@supp[['inner.radius']])
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

