% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{initStand}
\alias{initStand}
\title{Initialize the model Stand}
\usage{
initStand(area = 1000, npatch = 1, soil = c(0, -0.5, -1.5), z = 0,
  arrangement = "square", composition = "spatial", dist = 0.05)
}
\arguments{
\item{area}{patch size area}

\item{npatch}{number of patches}

\item{soil}{a vector or matrix of soil depths.}

\item{arrangement}{patch arrangement ('square' or 'linear').}

\item{composition}{'spatial' or 'temporal'}

\item{dist}{the fractional distance between the hexagons}
}
\value{
a \code{\link[DGVM3D]{Stand-class}}
}
\description{
Initialize the model Stand
}
\details{
If soil is a matrix, the number of columns must be equal to npatch. In that way each patch can have its own soil depth.
The patches represented as hexagons can either be arranged in a square or in a line. The later one for example to represent a time series (succession).
}
\examples{
stand <- initStand(npatch=9, z=rnorm(9))
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

