% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Patch-class}
\alias{Patch-class}
\title{One model patch}
\description{
This defines the basic class
}
\section{Slots}{

\describe{
\item{\code{id}}{unique ID}

\item{\code{pid}}{the patch id in the vegetation data.frame}

\item{\code{soil}}{vector of soil layer depth}

\item{\code{vegetation}}{the vegetation data.frame}

\item{\code{color.table}}{lookup table for coloring}
}}

