% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{random.disc}
\alias{random.disc}
\title{Random distribution in a circle}
\usage{
random.disc(n, strict = FALSE)
}
\arguments{
\item{n}{total number of trees}

\item{strict}{should the value 2pi be excluded}
}
\value{
data.frame of positions
}
\description{
Random distribution in a circle
}
