% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{row.disc}
\alias{row.disc}
\title{row-wise distribution of points in a disc}
\usage{
row.disc(n)
}
\arguments{
\item{n}{number of points}
}
\value{
data.frame of x and y positions
}
\description{
row-wise distribution of points in a disc
}
\examples{
par(mfrow=c(2,2), mai=c(0,0,0,0))
for (n in c(51, 250, 280, 1000)) {
  ret=row.disc(n)
  plot(sin(seq(0, 2*pi, length.out=361)), cos(seq(0, 2*pi, length.out=361)),
       type="l", axes = FALSE, ylab = "", xlab="")
  points(ret)
}
par(mfrow=c(2,2), mai=c(0,0,0,0))
for (n in c(120, 640, 1280, 2400)) {
  ret=row.disc(n)
  plot(sin(seq(0, 2*pi, length.out=361)), cos(seq(0, 2*pi, length.out=361)),
       type="l", axes = FALSE, ylab = "", xlab="")
  points(ret)
}
}
