% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{snapshot}
\alias{snapshot}
\title{Visualize a snapshot of patches.}
\usage{
snapshot(vegetation, stand.id = 1, patch.id = NULL, year = 2000)
}
\arguments{
\item{vegetation}{the data.frame of individual trees}

\item{stand.id}{the stand to take a snapshot off.}

\item{patch.id}{all patches (default) or just one.}

\item{year}{which year to take the snapshot off.}
}
\value{
a \code{\link{Stand-class}}.
}
\description{
Visualize a snapshot of patches.
}
\examples{
\dontrun{
stand=snapshot(dgvm3d.succession[[1]])
}
}
