% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Stand-class}
\alias{Stand-class}
\title{One model stand consisting of several patches}
\description{
One model stand consisting of several patches
}
\section{Slots}{

\describe{
\item{\code{patches}}{list of patches in one stand}

\item{\code{area}}{the area of each patch}

\item{\code{arrangement}}{either 'linear' or 'square'}
}}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

