% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{stand3D}
\alias{stand3D}
\title{3D view of the stands}
\usage{
stand3D(stand, patch.id = NULL)
}
\arguments{
\item{stand}{the \code{\link{Stand-class}} to visualize}

\item{patch.id}{the patch IDs to create. Default: all.}
}
\value{
None
}
\description{
Uses \code{\link{rgl}} to visualize a single, if patch.id is given, or all patch soil hexagons
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}
\seealso{
\code{\link{initStand}} for examples
}

