% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{sunflower.disc}
\alias{sunflower.disc}
\title{return the positions}
\usage{
sunflower.disc(n, alpha = 0)
}
\arguments{
\item{n}{total number of trees}

\item{alpha}{smoothing factor for boundary points}
}
\value{
position data.frame
}
\description{
return the positions
}
