% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{sunflower.radius}
\alias{sunflower.radius}
\title{Calculate the current radius}
\usage{
sunflower.radius(k, n, b)
}
\arguments{
\item{k}{current value}

\item{n}{total number}

\item{b}{number of boundary points}
}
\value{
radius
}
\description{
Calculate the current radius
}
\references{
\href{https://stackoverflow.com/questions/28567166/uniformly-distribute-x-points-inside-a-circle}{https://stackoverflow.com/questions/28567166/uniformly-distribute-x-points-inside-a-circle}
}
