% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{tree3D}
\alias{tree3D}
\title{draw a single tree}
\usage{
tree3D(tree = NULL, offset = c(0, 0, 0), col = c("#22BB22", "33FF33"),
  opacity = 1, faces = 29)
}
\arguments{
\item{tree}{one column of the \code{\link{Patch-class}} vegetation data.frame slot}

\item{offset}{x/y center and surface (z) of the respective patch}

\item{col}{crown colors for the shade classes}

\item{opacity}{alpha value for the tree crown (heavy impacting performance)}

\item{faces}{number of faces/triangles used per stem and tree cone}
}
\value{
None
}
\description{
draw a single tree
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

