% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TriangBody-class}
\alias{TriangBody-class}
\title{a Information to draw a triangular object}
\description{
a Information to draw a triangular object
}
\section{Slots}{

\describe{
\item{\code{vertices}}{the vertices of the object}

\item{\code{id}}{the column indices of the vertex matrix to draw the triangular body}

\item{\code{supp}}{supplementary information}
}}

