% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{updateStand}
\alias{updateStand}
\title{Remove/add trees with a new vegetation data.frame}
\usage{
updateStand(stand, vegetation, year = NULL)
}
\arguments{
\item{stand}{stand to update}

\item{vegetation}{new vegetation data.frame}

\item{year}{the next year}
}
\value{
stand
}
\description{
Removes those individuums with the shortest distance to any neighbour and adds new individuums randomly.
}

