% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGVM3D.R
\docType{package}
\name{DGVM3D-package}
\alias{DGVM3D}
\alias{DGVM3D-package}
\title{3D Visualization of simulated vegetation structure}
\description{
This is a visualization tool for simulated vegetation structure/succession in space and/or time of forest gap models. Those models should contain either individual trees or cohorts (e.g. LPJ-GUESS: Smith et al. (2001, 2014)).
}
\references{
Smith, B., Prentice, I. C. and Sykes, M. T.: Representation of vegetation dynamics in the modelling of terrestrial ecosystems: comparing two contrasting approaches within European climate space, Global Ecology and Biogeography, 10(6), 621-637, \href{https://dx.doi.org/10.1046/j.1466-822X.2001.t01-1-00256.x}{doi:10.1046/j.1466-822X.2001.t01-1-00256.x}, 2001.
Smith, B., Warlind, D., Arneth, A., Hickler, T., Leadley, P., Siltberg, J. and Zaehle, S.: Implications of incorporating N cycling and N limitations on primary production in an individual-based dynamic vegetation model, Biogeosciences, 11(7), 2027-2054, \href{https://dx.doi.org/10.5194/bg-11-2027-2014}{doi:10.5194/bg-11-2027-2014}, 2014.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joergsteinkamp/DGVM3D}
  \item Report bugs at \url{https://github.com/joergsteinkamp/DGVM3D/issues}
}

}
\author{
\strong{Maintainer}: Steinkamp Joerg \email{steinkamp.joerg@gmail.com}

}
