% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{dgvm3d.options}
\alias{dgvm3d.options}
\title{set some variables used in cascading functions}
\usage{
dgvm3d.options(x = NULL, patch.area = NULL, samples = NULL,
  overlap = NULL, sort.column = NULL, establish.method = NULL,
  color.column = NULL, verbose = NULL)
}
\arguments{
\item{x}{query character 'x' for its value.}

\item{patch.area}{the patch area in m^2.}

\item{samples}{2 element vector. 1. number of samples to determine the next trees position. 2. max. number to repeat the sampling}

\item{overlap}{fraction of crownradius allowed to overlap.}

\item{sort.column}{2 element vector: 1. vegetation data.frame column name to sort by. 2. "descending" (default) or "ascending".}

\item{establish.method}{where to place the trees: 'random', 'sunflower' or 'row'. If there are trees with positions already 'random' is applied.}

\item{color.column}{name of the vegetation column to create the canopy colors from.}

\item{verbose}{print some information.}
}
\description{
set some variables used in cascading functions
}
