% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{dgvm3d.options}
\alias{dgvm3d.options}
\title{set some variables used in cascading functions}
\usage{
dgvm3d.options(x = NULL, samples = NULL, overlap = NULL,
  sort.column = NULL, establish.method = NULL,
  establish.beta.parameters = NULL, verbose = NULL)
}
\arguments{
\item{x}{query character 'x' for its value.}

\item{samples}{2 element vector. 1. number of samples to determine the next trees position. 2. max. number to repeat the sampling}

\item{overlap}{fraction of crownradius allowed to overlap.}

\item{sort.column}{2 element vector: 1. vegetation data.frame culumn name to sort by. 2. "descending" (default) or "ascending".}

\item{establish.method}{where to place the next tree: 'random', 'min' or 'max' of valid sampled new positions.}

\item{establish.beta.parameters}{shape parameters for beta random value to get the distance from the patch center. This should be biased away from the center c(0.97, 1.4), otherwise trees tend to accumulate in the center.}

\item{verbose}{print some information.}
}
\description{
set some variables used in cascading functions
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

