% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{dgvm3d.succession}
\alias{dgvm3d.succession}
\title{timeseries data from 1865-2005 is 5 year steps}
\format{An object of class \code{data.frame} with 42619 rows and 19 columns.}
\usage{
dgvm3d.succession
}
\description{
@name dgvm3d.succession
 @docType data
 @author Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
 @keywords data
}
\keyword{datasets}

