% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{dgvm3d.succession}
\alias{dgvm3d.succession}
\title{Vegetation timeseries data from 1860-2005}
\format{An object of class \code{list} of length 3.}
\usage{
dgvm3d.succession
}
\description{
A list of 3 data.frames with simulation results of a LPJ-GUESS model run without random patch disturbance at 3 locations.
All patches at all locations were disturbed in 1859.
}
\details{
@name dgvm3d.succession
 @docType data
 @keywords data
}
\keyword{datasets}

