% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{stand3D}
\alias{stand3D}
\title{3D view of the stands}
\usage{
stand3D(stand, patch.id = NULL)
}
\arguments{
\item{stand}{the \code{\link[DGVM3D]{Stand-class}} to visualize}
}
\description{
3D view of the stands
}
\author{
Joerg Steinkamp \email{steinkamp.joerg@gmail.com}
}

