% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{sunflower.radius}
\alias{sunflower.radius}
\title{Calc the current radius}
\usage{
sunflower.radius(k, n, b)
}
\arguments{
\item{k}{current value}

\item{n}{total number}

\item{b}{number of boundary points}
}
\value{
radius
}
\description{
Calc the current radius
}

